TIDIJ-20 Appointment display in real time
